fix: db connections worker (#1714)
